Fixed way of determing the register from the numer in th opcode
Previously the disassembler was using the dirty trick to sum one to the number
of the register to determine which register was. That worked with the previous
version of the backend but now tablegen places SR register before all the
others. To avoid further dirty fixes it has been implemented in a different
way.